* Zone.h: Use size_t instead of unsigned for return of getPageSize.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@3627 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/Zone.h
@@ -29,7 +29,7 @@ Library:      defobj
 + createBegin: aZone;
 - (void)setPageSize: (size_t)pageSize;
 - createEnd;
-- (unsigned)getPageSize;
+- (size_t)getPageSize;
 - allocIVars: aClass;
 - copyIVars: anObject;
 - (void)freeIVars: anObject;
